+++ dependencies/FSParam/FSParam.csproj
@@ -10,8 +10,4 @@
     <ProjectReference Include="..\SoulsFormats\SoulsFormats\SoulsFormats.csproj" />
     <ProjectReference Include="..\StudioUtils\StudioUtils.csproj" />
   </ItemGroup>
-    <!-- <ItemGroup>
-        <PackageReference Include="BouncyCastle.Cryptography" Version="2.3.1" />
-        <PackageReference Include="SharpZipLib" Version="1.4.2" />
-    </ItemGroup> -->
 </Project>
