+++ src/ERBingoRandomizer/Tasks/Randomizer.Equipment.cs
@@ -1,11 +1,9 @@
 ﻿using Project.Params;
 using System;
 using System.Collections.Generic;
-using System.Collections.Specialized;
 using System.Linq;
 using Project.Settings;
 using FSParam;
-using Project.Utility;
 
 namespace Project.Tasks;
